tests/libostreetest: check $PWD not / for overlay
This was originally part of #1199, but never got in. The work in #1218
is awesome for CI, though for the developer workflow, I find it much
easier to just set $TEST_TMPDIR to a non-overlayfs mount point. Teach
the test libraries to just look at $PWD rather than /.

Also export the function that does the overlayfs checking. Prep for a
future patch.